Utilisation de VehicleInformation
git-tfs-id: [http://172.17.236.41:8080/]$/AdExpress International Demo/Trunk;C4912

+++ DLL/TNS.AdExpress.Web.Controls/Results/Creatives/CreativesHeaderWebControl.cs
@@ -58,11 +58,11 @@ namespace TNS.AdExpress.Web.Controls.Results.Creatives {
         /// <summary>
         /// Vehicle List
         /// </summary>
-        protected IList<int> _vehicles = new List<int>();
+		protected IList<long> _vehicles = new List<long>();
         /// <summary>
         /// Get / Set list of vehicles
         /// </summary>
-        public IList<int> Vehicles {
+		public IList<long> Vehicles {
             get {
                 return _vehicles;
             }
@@ -73,11 +73,11 @@ namespace TNS.AdExpress.Web.Controls.Results.Creatives {
         /// <summary>
         /// Cuurent Vehicle Id
         /// </summary>
-        protected int _idVehicle = -1;
+		protected long _idVehicle = -1;
         /// <summary>
         /// Get / Set current vehicle id
         /// </summary>
-        public int IdVehicle {
+		public long IdVehicle {
             get {
                 return _idVehicle;
             }
